#minor update user management
